+++ internal/sql_adapter.go
@@ -24,8 +24,8 @@ func (a *SqlAdapter) RowName() string {
 	return "row"
 }
 
-func (a *SqlAdapter) Scan(urlStr string, showData bool, showAll bool, limit int, matchConfig *MatchConfig) ([]ruleMatch, error) {
-	return scanDataStore(a, urlStr, showData, showAll, limit, matchConfig)
+func (a *SqlAdapter) Scan(scanOpts ScanOpts) ([]ruleMatch, error) {
+	return scanDataStore(a, scanOpts)
 }
 
 func (a *SqlAdapter) Init(url string) error {